attacker/requirements.txt
qrcode[pil]
pillow